introduce automatic versioning via git
Note: does not work with e.g. IntelliJ IDEA, because it builds the
project out of the git directory! In this case 'empty version name' is
shown as version.

Get the version number from an annotated git tag (parameter '-a').

  versionName: ignore leading non-digits and everything after the first underscore
  versionCode: take digits after the first underscore

To get versionName '0.1' and versionCode '2' valid tags are, e.g. 0.1_2,
release-0.1_2, foo0.1_2, foo0.1_2_bar. See source code comments for
further information.

If no tag is found, the current branch is taken as versionName. Default
versionCode is '1'.

Source: http://orip.org/2013/10/versioning-android-apps-with-git-tags.html