add gastro contact details
This commit fills the gastro details fragment with content. It adds the
contact information: The address as clickable google maps link and the
telephone number, which is also clickable and gets a user to his
favorite telephone app.

The telephone number was missing from the gastro location, which has
been fixed in this commit as well.