do not recreate main list activity again and again
When the back button is pressed, e.g. in settings activity or in gastro
activity, main list activity's onCreate is called. So the json file is
fetched again from the server and the location is retrieved again as
well. For the latter we have introduced a swipe refresh layout. We
should avoid downloading the json file again and again, while navigating
through the application.

Source: http://stackoverflow.com/questions/12276027/how-can-i-return-to-a-parent-activity-correctly/15933890#15933890