simplify release versioning
* version name is the annotated tag found on a commit or the branch you
are currently on

* version code has to be set manually for the following reason: until
today the version code would be increased with every release, because of
the tag, e.g. '0.1_1', '0.1_2', '0.1_3', and so on. After a release it
would be directly set to '1' again, because of the missing tag. That way
Android would complain after every release, that a developer wants to
install an old version of the app. The version name is not taken into
account by Android in this case