improve layout with help of keyline pushing
Keyline Pushing v1.2 by Faiz Malkani is an app, which helps a developer
to check if the app fulfills the Google style guidelines for material
design. It draws an 8dp grid and the keyline over all other app,
allowing one to easily check the baselines of any app's components.

Also set card corner radius to 0dp to have the rectangular material
design.

Note: this is still not the final layout.