remove soft hyphens from gastro description
Unfortunately soft hyphen (&shy;) is only partially working with
fromHtml(): the word gets split at the correct place, but the hyphen
(dash) is not shown. This might be very annoying to the user, because it
just does not look right. As a workaround we do not split words at all.

Closes #13