introduce gastro locations class
Introduce a class, that holds an object of all gastro locations,
filtered gastro locations, and the gastro locations, which are currently
presented to the user. Latter includes searching through the gastro
locations lists.

This shrinks the main list activity and centralizes functionality used
on the gastro locations in one object.