remove hard coded acra shared secret
We had some spammers lately. So we pass the shared secret for ACRA to
the gradle build system via '-PsharedSecret=<acra_shared_secret>'.

Note: this is only relevant to release builds, since acra is disabled
for debug builds. A developer does not have to add a parameter to
his/her gradle config. The shared secret is just empty then.