fix expandable text view on orientation change
Without this patch the expandable text view had the following behavior
on orientation change while in collapsed state: from portrait to
landscape the text would fit into less than three lines, from landscape
to portrait the ellipsis would not show.

This happened because the text, that would fit into the current layout,
was not re-calculated.