be consistent regarding the nav drawer
At start of our app the gastronomy locations are displayed. Pulling out
the navigation drawer from the left has no selection. Clicking on
'gastronomy' the item is selected in the drawer and the view does not
change, because it is already displayed.

Setting the default selected item to 'gastronomy' at start up makes the
app behaviour more consistent.