Move map overview activity to a frament
The problem of this commit: It is not working at the moment :(.

The fragments on their own are working (more or less perfectly)
if you add them right in the beginning (at line 135 of
LocationListActivity.java), but the swap is not working.

Perhaps it is not possible to do a fragment replace within a SwipeRefreshLayout?

FYI: The map fragment doesn't show any entries if you
add it at the beginning, but it does show the entries
if you add *no* fragment at the beginning and add
the map fragment on icon press in the menu bar.

Just uncomment

>         getSupportFragmentManager().beginTransaction().add(mSwipeRefreshLayout.getId(),
>         mLocationListFragment).commit();

in line 135 of LocationListActivity.java, compile&run, hit the
map button (the earth in the top bar), and the map works.