Fix fragments of list view/map overview
Before this patch, we were able to switch between list view and map
overview, but the map view was empty, because of the swipe refresh
layout. This layout is not good to use with a map view anyways, because
scrolling down triggers the refresh.

We therefore use just a frame layout around the swipe layout like
Android suggests in the documentation.

Further changes:
* use correct map icon from http://github.com/google/material-design-icons/
* add list icon from http://github.com/google/material-design-icons/
* make the menu bar consistent: in the map overview just display an
  entry to switch back. Also if the user searched in the list view,
  dismiss the search (mSearchItem.collapseActionView()). We do not
  support a dynamic map overview currently