Remove product flavors
We decided to ditch the PlayStore build variant and only support FOSS,
because the maintenance got too high.

Maybe we need the product flavors for distinguishing between FOSS and
PlayStore app rating. This needs further investigation first. One of the
F-Droid guys mentioned, it would be no problem to have the PlayStore app
rating in the FOSS variant, so we are good for now.